requirements.txt
sqlalchemy
mariadb==1.1.2
fastapi[all]
pyjwt